refactor: LineService에 개선된 Line 도메인 객체 적용
- 테스트 리팩터링
- Section 빠진 검증 추가
- LineDto 추가
- DTO 패키지 이동
