+++ README2.md
@@ -35,6 +35,8 @@
 - 구간
   - 두 역이 모두 있어야 한다.
   - 어떤 구간과 연결되는지 알 수 있다. 
+    - [ ] 서로 대칭인 경우는 연결된 경우가 아니다.   
+      (`잠실나루` - `잠실`, `잠실` - `잠실나루`)
   - 어떤 구간과 겹치는지 알 수 있다.
   - 어떤 역을 포함하는지 알 수 있다.
   - 한 구간을 일부분들로 쪼갠다.
