+++ src/main/java/subway/application/LineService.java
@@ -1,148 +1,119 @@
 package subway.application;
 
-import org.springframework.stereotype.Service;
-import subway.dao.LineDao;
-import subway.dao.SectionDao;
-import subway.dao.StationDao;
-import subway.domain.Line;
-import subway.domain.Station;
-import subway.dto.*;
-
-import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
+
+import org.springframework.stereotype.Service;
+
+import subway.dao.LineRepository;
+import subway.dao.StationDao;
+import subway.domain.Distance;
+import subway.domain.Line;
+import subway.domain.Section;
+import subway.domain.Station;
+import subway.dto.LineRequest;
+import subway.dto.LineResponse;
+import subway.dto.SectionRequest;
+import subway.dto.StationDeleteRequest;
+import subway.dto.StationRegisterRequest;
+import subway.dto.StationsRegisterRequest;
 
 @Service
 public class LineService {
-    private final LineDao lineDao;
-    private final StationDao stationDao;
-    private final SectionDao sectionDao;
 
-    public LineService(LineDao lineDao, StationDao stationDao, SectionDao sectionDao) {
-        this.lineDao = lineDao;
+    private final StationDao stationDao;
+
+    private final LineRepository lineRepository;
+
+    public LineService(StationDao stationDao, LineRepository lineRepository) {
         this.stationDao = stationDao;
-        this.sectionDao = sectionDao;
+        this.lineRepository = lineRepository;
     }
 
     public LineResponse saveLine(LineRequest request) {
-        Line persistLine = lineDao.insert(new Line(request.getName(), request.getColor()));
-        return LineResponse.of(persistLine);
+        Line line = new Line(request.getName(), request.getColor());
+        Line saved = lineRepository.save(line);
+        return LineResponse.of(saved);
     }
 
     public List<LineResponse> findLineResponses() {
-        List<Line> persistLines = findLines();
-        return persistLines.stream()
+        List<Line> lines = lineRepository.findAll();
+        return lines.stream()
                 .map(LineResponse::of)
                 .collect(Collectors.toList());
     }
 
-    public List<Line> findLines() {
-        return lineDao.findAll()
-                .stream()
-                .map(line -> findLineById(line.getId()))
-                .collect(Collectors.toList());
-    }
-
     public LineResponse findLineResponseById(Long id) {
-        Line persistLine = findLineById(id);
-        return LineResponse.of(persistLine);
+        Line found = lineRepository.findBy(id);
+        return LineResponse.of(found);
     }
 
-    public Line findLineById(Long id) {
-        Line rawLine = lineDao.findById(id);
-        List<SectionDto> sections = sectionDao.findAllByLineId(id);
-        if (sections.isEmpty()) {
-            return rawLine;
-        }
+    /* 학습용 주석입니다 :)
+    DAO를 직접 사용해 DB에 업데이트하기 vs 도메인에 setter 넣기
+    후자를 선택했다.
+    DAO를 사용한다면 캡슐화를 지킬 순 있지만 서비스 레이어에서 저장 방법을 신경써야 한다.
+    관심사가 섞여서 복잡도가 올라가고, 유지보수성과 유연성이 떨어진다. 이는 Repository를 만든 이유와 정반대된다.
+    더불어 사용자가 정말 이름을 바꾸길 원한다면, 못 바꿀 이유는 또 뭔가?
 
-        Map<Map.Entry<Station, Station>, Integer> distances = new HashMap<>();
-        sections.forEach(section ->
-            distances.put(Map.entry(stationDao.findById(section.getStationId()), stationDao.findById(section.getNextStationId())), section.getDistance()));
+    이름은 서비스 영역에서 마음대로 바꿔도 되는가?
+    이름이 중복되면 안된다고 가정해보자.
+    겹치는 이름이 있다면, 그건 이미 저장된 이름들이다.
+    따라서 저장 영역에서 그런 이름이 가능한지 검증이 있으면 되지 않을까?
+    그럼 관심사가 분리된 구조를 지키면서도 안전하게 요청을 처리할 수 있다.
+    */
+    public void updateLine(Long id, LineRequest lineRequest) {
+        Line line = lineRepository.findBy(id);
 
-        SectionDto section = sections.remove(0);
-        LinkedList<Long> stationIds = new LinkedList<>(List.of(section.getStationId(), section.getNextStationId()));
+        line.setName(lineRequest.getName());
+        line.setColor(lineRequest.getColor());
 
-        while(!sections.isEmpty()) {
-            for (int i = 0; i < sections.size(); i++) {
-                section = sections.get(i);
-                if (section.getStationId().equals(stationIds.getLast())) {
-                    stationIds.addLast(section.getNextStationId());
-                    sections.remove(i);
-                    break;
-                }
-                if (section.getNextStationId().equals(stationIds.getFirst())) {
-                    stationIds.addFirst(section.getStationId());
-                    sections.remove(i);
-                    break;
-                }
-            }
-        }
-
-        List<Station> stations = stationIds.stream()
-                .map(stationDao::findById)
-                .collect(Collectors.toList());
-
-        return new Line(rawLine.getId(), rawLine.getName(), rawLine.getColor(), new LinkedList<>(stations), distances);
-    }
-
-    public void updateLine(Long id, LineRequest lineUpdateRequest) {
-        lineDao.update(new Line(id, lineUpdateRequest.getName(), lineUpdateRequest.getColor()));
+        lineRepository.save(line);
     }
 
     public void deleteLineById(Long id) {
-        lineDao.deleteById(id);
+        Line line = lineRepository.findBy(id);
+        lineRepository.delete(line);
     }
 
     public void registerStations(Long id, StationsRegisterRequest request) {
-        Line line = findLineById(id);
-        Station top = stationDao.findById(request.getTopId());
-        Station bottom = stationDao.findById(request.getBottomId());
+        Line line = lineRepository.findBy(id);
+        Station upper = stationDao.findById(request.getTopId());
+        Station lower = stationDao.findById(request.getBottomId());
 
-        line.insertBoth(top, bottom, request.getDistance());
+        Section section = new Section(upper, lower, request.getDistance());
+        line.add(section);
 
-        sectionDao.insertAll(toSectionDtos(line));
+        lineRepository.save(line);
     }
 
     public void registerStation(Long id, StationRegisterRequest request) {
-        Line line = findLineById(id);
-        Station station = stationDao.findById(request.getStationId());
-        Station base = stationDao.findById(request.getBaseId());
-        String where = request.getWhere();
+        Line line = lineRepository.findBy(id);
+        Station upper = stationDao.findById(request.getStationId());
+        Station lower = stationDao.findById(request.getBaseId());
 
-        if (where.equals("UPPER")) {
-            line.insertUpper(station, base, request.getDistance());
-        }
+        line.add(new Section(upper, lower, request.getDistance()));
 
-        if (where.equals("LOWER")) {
-            line.insertLower(station, base, request.getDistance());
-        }
+        lineRepository.save(line);
+    }
 
-        sectionDao.deleteAllByLineId(line.getId());
-        sectionDao.insertAll(toSectionDtos(line));
+    public void addSection(Long lineId, SectionRequest sectionRequest) {
+        Line line = lineRepository.findBy(lineId);
+        Station upper = stationDao.findById(sectionRequest.getUpperStationId());
+        Station lower = stationDao.findById(sectionRequest.getLowerStationId());
+
+        Distance distance = new Distance(sectionRequest.getDistance());
+        Section section = new Section(upper, lower, distance.getValue());
+        line.add(section);
+
+        lineRepository.save(line);
     }
 
     public void deleteStation(Long lineId, StationDeleteRequest request) {
-        Line line = findLineById(lineId);
+        Line line = lineRepository.findBy(lineId);
         Station station = stationDao.findById(request.getStationId());
-        line.delete(station);
 
-        sectionDao.deleteAllByLineId(line.getId());
-        sectionDao.insertAll(toSectionDtos(line));
-    }
-    //현재 상행, 하행 모든 내용을 저장하고 있음
+        line.remove(station);
 
-    private List<SectionDto> toSectionDtos(Line line) {
-        return line.getAdjacentStations().stream()
-                .map(entry -> toSectionDto(
-                        line,
-                        entry.getKey(),
-                        entry.getValue()))
-                .collect(Collectors.toList());
-    }
-
-    private SectionDto toSectionDto(Line line, Station station, Station nextStation) {
-        return new SectionDto(line.getId(), station.getId(), nextStation.getId(), line.getDistanceBetween(station, nextStation));
+        lineRepository.save(line);
     }
 }
