+++ src/test/java/subway/dao/SectionDaoTest.java
@@ -12,7 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 
-import subway.dto.SectionDto;
+import subway.dao.dto.SectionDto;
 
 @JdbcTest
 public class SectionDaoTest {
