+++ src/test/java/subway/dao/SectionRepositoryTest.java
@@ -25,7 +25,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import subway.domain.Section;
-import subway.dto.SectionDto;
+import subway.dao.dto.SectionDto;
 
 @SuppressWarnings("NonAsciiCharacters")
 @ExtendWith(MockitoExtension.class)
